requirements.txt
beautifulsoup4==4.14.3
pandas==3.0.1
psycopg2-binary==2.9.11
python-dotenv==1.2.2
PyYAML==6.0.1
selenium==4.41.0
SQLAlchemy==2.0.48
